+++ vitest.config.ts
@@ -4,10 +4,10 @@ export default defineConfig({
   test: {
     include: ['./test/tests/**/*.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
     coverage: {
-      statements: 91,
-      branches: 78,
+      statements: 89,
+      branches: 77,
       functions: 87,
-      lines: 91,
+      lines: 89,
       exclude: ['test', 'src/index.ts'],
     },
   },