+++ CHANGELOG.md
@@ -12,6 +12,15 @@ Unreleased changes are in the `master` branch.
 ### Added
 
 - `/utils/txs/evaluate/utxos` method
+- initial support for POST endpoints to `getSchemaForEndpoint`
+
+### Changed
+
+- Migrated to OpenAPI 3.1.0
+
+### Removed
+
+- named error responses from OpenAPI spec (overusage_limit, mempool_full, pin_queue_full, autobanned, not_found, internal_server_error, unauthorized_error, bad_request).
 
 ### Fixed